Offer choice between discarding and switching to zeros negative adjusted counts
Following Rothman "Treatment of Negative Adjustments", this gives the
option to discard (i.e. replace by NA as in episensr 0.6.0) or replace
by zeros the negative adjusted counts. It also gives warning and
conventional messages if all/too many negative values replaced. It also
simplyfies the print statement and include a note on which discard
method was chosen.